requestapp/requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
freeze==3.0
Pillow==8.0.1
psycopg2==2.8.6
python-memcached==1.59
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
